Rename renderers to InputRenderer
